♻️ 코드 수정 : 댓글 삭제 기능 수정 (컨트롤러 - Dto 사용, 서비스 - Dto 활용 유저 인증 추가)

+++ src/main/java/bigsanghyuk/four_uni/comment/service/CommentService.java
@@ -1,5 +1,6 @@
 package bigsanghyuk.four_uni.comment.service;
 
+import bigsanghyuk.four_uni.comment.domain.DeleteCommentInfo;
 import bigsanghyuk.four_uni.comment.domain.EditCommentInfo;
 import bigsanghyuk.four_uni.comment.domain.RegisterCommentInfo;
 import bigsanghyuk.four_uni.comment.domain.entity.Comment;
@@ -7,6 +8,7 @@ import bigsanghyuk.four_uni.comment.repository.CommentRepository;
 import bigsanghyuk.four_uni.comment.repository.LikeCommentRepository;
 import bigsanghyuk.four_uni.exception.comment.CommentEditOtherUserException;
 import bigsanghyuk.four_uni.exception.comment.CommentNotFoundException;
+import bigsanghyuk.four_uni.exception.comment.CommentRemoveOtherUserException;
 import bigsanghyuk.four_uni.exception.post.PostNotFoundException;
 import bigsanghyuk.four_uni.exception.user.UserNotFoundException;
 import bigsanghyuk.four_uni.post.repository.PostRepository;
@@ -60,9 +62,19 @@ public class CommentService {
     }
 
     @Transactional
-    public void remove(Long commentId) {
-        commentRepository.findById(commentId)
+    public void remove(Long commentId, @Valid DeleteCommentInfo deleteCommentInfo) {
+        postRepository.findById(deleteCommentInfo.getPostId())
+                .orElseThrow(PostNotFoundException::new);
+
+        commentRepository.findByUserIdOrderByIdDesc(deleteCommentInfo.getUserId())
+                .orElseThrow(UserNotFoundException::new);
+
+        Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(CommentNotFoundException::new);
+
+        if (!deleteCommentInfo.getUserId().equals(comment.getUserId())) {
+            throw new CommentRemoveOtherUserException();
+        }
 
         commentRepository.deleteCommentByAndId(commentId);
         likeCommentRepository.deleteLikeCommentByCommentId(commentId);
